add ProtectedRoute
